// src/models/entity/system_info_entity.ts
/*
 * --------------------------------------------------------------------------------
 * Copyright (c) 2016-NOW(至今) 筱锋
 * Author: 筱锋(https://www.x-lf.com)
 *
 * 本文件包含 XiaoMain 的源代码，该项目的所有源代码均遵循MIT开源许可证协议。
 * --------------------------------------------------------------------------------
 * 许可证声明：
 *
 * 版权所有 (c) 2016-2024 筱锋。保留所有权利。
 *
 * 本软件是“按原样”提供的，没有任何形式的明示或暗示的保证，包括但不限于
 * 对适销性、特定用途的适用性和非侵权性的暗示保证。在任何情况下，
 * 作者或版权持有人均不承担因软件或软件的使用或其他交易而产生的、
 * 由此引起的或以任何方式与此软件有关的任何索赔、损害或其他责任。
 *
 * 使用本软件即表示您了解此声明并同意其条款。
 *
 * 有关MIT许可证的更多信息，请查看项目根目录下的LICENSE文件或访问：
 * https://opensource.org/licenses/MIT
 * --------------------------------------------------------------------------------
 * 免责声明：
 *
 * 使用本软件的风险由用户自担。作者或版权持有人在法律允许的最大范围内，
 * 对因使用本软件内容而导致的任何直接或间接的损失不承担任何责任。
 * --------------------------------------------------------------------------------
 */


/**
 * # SystemInfoEntity 接口
 *
 * 表示系统信息的数据结构接口。包括网站信息和博主信息的详细描述。
 * 该接口用于定义系统的基础配置信息。
 *
 * ## 属性说明
 * - `site`: 包含网站的基本信息。
 * - `blogger`: 包含博主的详细信息。
 */
export interface SystemInfoEntity {
    /**
     * 网站信息
     * @param site_name 网站名称
     * @param author 网站作者
     * @param version 网站版本
     * @param description 网站描述
     * @param keywords 网站关键词
     */
    site: {
        site_name: string;
        author: string;
        version: string;
        description: string;
        keywords: string;
    };

    /**
     * 博主信息
     * @param name 博主姓名
     * @param nick 博主昵称
     * @param email 博主邮箱
     * @param description 博主描述
     */
    blogger: {
        name: string;
        nick: string;
        email: string;
        description: string;
    };
}